+++ backend/src/main/java/com/rssmanager/auth/service/GithubSessionAuthService.java
@@ -3,8 +3,8 @@ package com.rssmanager.auth.service;
 import com.rssmanager.auth.controller.dto.CertificateResponse;
 import com.rssmanager.auth.controller.dto.LoginRequest;
 import com.rssmanager.auth.controller.dto.LoginResponse;
+import com.rssmanager.member.application.MemberService;
 import com.rssmanager.member.controller.dto.MemberResponse;
-import com.rssmanager.member.service.MemberService;
 import com.rssmanager.util.SessionManager;
 import java.util.Objects;
 import org.springframework.stereotype.Service;
